move PhpDocInfo unde PhpDocInfo namespace
