+++ packages/BetterPhpDocParser/src/DependencyInjection/CompilerPass/CollectDecoratorsToPhpDocInfoFactoryCompilerPass.php
@@ -5,7 +5,7 @@ namespace Symplify\BetterPhpDocParser\DependencyInjection\CompilerPass;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symplify\BetterPhpDocParser\Contract\PhpDocInfoDecoratorInterface;
-use Symplify\BetterPhpDocParser\PhpDocParser\PhpDocInfoFactory;
+use Symplify\BetterPhpDocParser\PhpDocInfo\PhpDocInfoFactory;
 use Symplify\PackageBuilder\DependencyInjection\DefinitionCollector;
 use Symplify\PackageBuilder\DependencyInjection\DefinitionFinder;
