+++ packages/BetterPhpDocParser/src/PhpDocInfo/PhpDocInfoFactory.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Symplify\BetterPhpDocParser\PhpDocParser;
+namespace Symplify\BetterPhpDocParser\PhpDocInfo;
 
 use PHPStan\PhpDocParser\Lexer\Lexer;
 use PHPStan\PhpDocParser\Parser\PhpDocParser;
