+++ packages/BetterPhpDocParser/tests/PhpDocModifier/PhpDocModifierTest.php
@@ -3,7 +3,7 @@
 namespace Symplify\BetterPhpDocParser\Tests\PhpDocModifier;
 
 use Iterator;
-use Symplify\BetterPhpDocParser\PhpDocParser\PhpDocInfoFactory;
+use Symplify\BetterPhpDocParser\PhpDocInfo\PhpDocInfoFactory;
 use Symplify\BetterPhpDocParser\Printer\PhpDocInfoPrinter;
 use Symplify\BetterPhpDocParser\Tests\AbstractContainerAwareTestCase;
